Feat: Gifticon 에 대한 테이블 추가 및 history 테이블 추가

Closes #29

See merge request s12-blockchain-nft-sub1/S12P21E101!30

+++ backend/src/main/java/com/e101/nift/article/model/dto/request/ArticleLikeDTO.java
@@ -13,5 +13,5 @@ public class ArticleLikeDTO {
     private String title;
     private int countLikes;
     private String imageUrl;
-    private int currentPrice;
+    private Float currentPrice;
 }
